Move SubSectionDivider to flat structure
