+++ packages/components/src/components/SubSectionDivider/index.ts
@@ -9,11 +9,4 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './InputField/InputField'
-export * from './InputField/InputError'
-export * from './TextInput'
-export * from './RadioGroup/RadioGroup'
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './PDFViewer/PDFViewer'
-export * from './InputField/InputLabel'
+export * from './SubSectionDivider'
